ti: retimer: i2c access in critical section
Summary:
Add lock for retimer channel access using i2c. Channel selection and i2c
register access must be done under lock to prevent multi-client r/w
operations.
Add reset all chan sysfs entry

Ref T17884

Test Plan: eth_loopback

Reviewers: O39 ethernet!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T17884

Differential Revision: https://phab.kalray.eu/D17867

+++ drivers/net/ethernet/kalray/kvx-net-hw.h
@@ -827,7 +827,7 @@ struct kvx_eth_hw {
 	u32 aggregated_only;
 	u32 eth_id;
 	u64 mppa_id;
-	u16 dev_id;
+	u64 dev_id;
 	struct pll_cfg pll_cfg;
 	u32 asn;
 	u32 vchan;
